+++ CHANGES.md
@@ -4,6 +4,10 @@ This files lists the changes during the lifetime of this project.
 
 ## unreleased
 
+- 220: support multiple organizations per assignment
+- 220: (migration) add AssignmentOrganizationUnit.role (PRIMARY, INVOLVED)
+- 220: make dummy data correct
+
 ## 2026-03-06
 
 - add endpoint to worker container for health checks